fix file:// URLs on Windows
BUG In Node.js on Windows, file: URLs have three slashes before the
drive letter. For example:

    file:///c:/Users/strager/[snip]/nsfwjs/model.json

When converting a file: URL to a file path, NodeFileSystem's router
incorrectly preserves one leading slash. This results in a
drive-relative path, which refers to a different file than intended.
This leads to errors down the road:

    Error: Path C:\C:\Users\strager\[snip]\nsfwjs\model.json does not exist: loading failed
        at C:\Users\strager\[snip]\node_modules\@tensorflow\tfjs-node\dist\io\file_system.js:67:23

Use Node.js' built-in functions for dealing with file: URLs instead of
hackily parsing them ourselves.

On POSIX systems, this commit should not change behavior.